# swagger.yaml
swagger: '2.0'
info:
  description: ViruSafe Backend REST API
  version: 1.0.0
  title: ViruSafe Backend REST API
host: 'localhost:8080'
basePath: /
...
paths:
...
  /questionnaire:
    get:
      tags:
        - questionnaire-controller
      summary: Get all questionnaire questions
      operationId: getQuestionnaireUsingGET
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          type: string
          x-example: Bearer eyJhbGciOiJIUzI1NiJ9.eyJjcmVhdGVkT24iOiIxNTY3NjUyNDAwIiwidXNlckd1aWQiOiJiYTBkNDdhYS0yYjYwLTQ5OTAtOGU5MS04MDIwYWU3ZTRkMTAiLCJqd3RTZWNyZXQiOiI3NjFiOGEyNi03N2YxLTQxNDEtOWVjNy0wZmZkMmEwZThkNjkiLCJwaG9uZU51bWJlciI6IjU1NTU1NTU1NTUiLCJleHAiOjE1OTkyNzQ4MDB9.DHgcQA0NAa7nySMD02dhhQKrcYP7LvC9AgJYLp-S0nQ
        - name: clientId
          in: header
          description: Client ID
          required: true
          type: string
          x-example: 3b3d3b02-e052-4289-827b-6c2f2e2a0430
        - name: language
          in: header
          description: Language
          required: false
          type: string
          x-example: en
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/QuestionDTO'
            example:
              - "id": 1
                "questionTitle": "Body temperature above 38C"
                "questionType": "BOOLEAN"
              - "id": 2
                "questionTitle": "Dry cough"
                "questionType": "BOOLEAN"
              - "id": 3
                "questionTitle": "Difficulty breathing or shortness of breath"
                "questionType": "BOOLEAN"
              - "id": 4
                "questionTitle": "Persistent pain or pressure in the chest"
                "questionType": "BOOLEAN"
              - "id": 5
                "questionTitle": "Bluish lips or face"
                "questionType": "BOOLEAN"
              - "id": 6
                "questionTitle": "Nausea, vomiting and/or diarrhoea"
                "questionType": "BOOLEAN"
              - "id": 7
                "questionTitle": "Lightheadedness or Vertigo"
                "questionType": "BOOLEAN"
        '401':
          description: Access unauthorized
          schema:
            $ref: '#/definitions/ErrorDTO'
        '403':
          description: Access forbidden
          schema:
            $ref: '#/definitions/ErrorDTO'
        '404':
          description: Not Found
        '500':
          description: An unexpected server error has occurred
          schema:
            $ref: '#/definitions/ErrorDTO'
      deprecated: false
    post:
      tags:
        - questionnaire-controller
      summary: Submit a user's questionnaire actions
      operationId: postQuestionnaireUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          type: string
          x-example: Bearer eyJhbGciOiJIUzI1NiJ9.eyJjcmVhdGVkT24iOiIxNTY3NjUyNDAwIiwidXNlckd1aWQiOiJiYTBkNDdhYS0yYjYwLTQ5OTAtOGU5MS04MDIwYWU3ZTRkMTAiLCJqd3RTZWNyZXQiOiI3NjFiOGEyNi03N2YxLTQxNDEtOWVjNy0wZmZkMmEwZThkNjkiLCJwaG9uZU51bWJlciI6IjU1NTU1NTU1NTUiLCJleHAiOjE1OTkyNzQ4MDB9.DHgcQA0NAa7nySMD02dhhQKrcYP7LvC9AgJYLp-S0nQ
        - name: clientId
          in: header
          description: Client ID
          required: true
          type: string
          x-example: 3b3d3b02-e052-4289-827b-6c2f2e2a0430
        - in: body
          name: questionnairePostDTO
          description: questionnairePostDTO
          required: true
          schema:
            $ref: '#/definitions/QuestionnairePostDTO'
      responses:
        '201':
          description: Successfully submitted questionnaire
        '401':
          description: Access unauthorized
          schema:
            $ref: '#/definitions/ErrorDTO'
        '403':
          description: Access forbidden
          schema:
            $ref: '#/definitions/ErrorDTO'
        '404':
          description: No corresponding user was found
          schema:
            $ref: '#/definitions/ErrorDTO'
        '412':
          description: Request argument validation has failed
          schema:
            $ref: '#/definitions/ValidationErrorDTO'
        '429':
          description: Request rate limit has been exceeded
          schema:
            $ref: '#/definitions/ErrorDTO'
        '500':
          description: An unexpected server error has occurred
          schema:
            $ref: '#/definitions/ErrorDTO'
      deprecated: false
...
definitions:
...
  AnswerDTO:
    type: object
    properties:
      answer:
        type: string
        example: 'TRUE'
        pattern: '^[a-zA-Z0-9 ,.()а-яА-Я-]+$'
      questionId:
        type: integer
        format: int32
        example: 3
    title: AnswerDTO
  ErrorDTO:
    type: object
    properties:
      message:
        type: string
        example: Invalid request!
    title: ErrorDTO
  FieldValidationErrorDTO:
    type: object
    properties:
      fieldName:
        type: string
    title: FieldValidationErrorDTO
  Location:
    type: object
    properties:
      lat:
        type: number
        format: double
        example: 42.71
        minimum: -90
        maximum: 90
        exclusiveMinimum: false
        exclusiveMaximum: false
      lng:
        type: number
        format: double
        example: 23.32
        minimum: -180
        maximum: 180
        exclusiveMinimum: false
        exclusiveMaximum: false
    title: Location
...
  QuestionDTO:
    type: object
    properties:
      id:
        type: integer
        format: int32
        example: 3
      questionTitle:
        type: string
        example: Difficulty breathing or shortness of breath
      questionType:
        type: string
        example: BOOLEAN
        enum:
          - BOOLEAN
          - OPEN_ANSWER
    title: QuestionDTO
  QuestionnairePostDTO:
    type: object
    properties:
      answers:
        type: array
        items:
          $ref: '#/definitions/AnswerDTO'
      location:
        $ref: '#/definitions/Location'
      timestamp:
        type: integer
        format: int64
        example: 1567652400
    title: QuestionnairePostDTO
...
  ValidationErrorDTO:
    type: object
    properties:
      message:
        type: string
        example: Request arguments validation failed!
      validationErrors:
        type: array
        items:
          $ref: '#/definitions/FieldValidationErrorDTO'
    title: ValidationErrorDTO